refactor(composer): move the composer into its own folder
