+++ desk/src/components/communication-area/CommunicationArea.vue
