+++ desk/src/components/index.ts
@@ -2,7 +2,7 @@ export { default as BackButton } from "./BackButton.vue";
 export { default as AttachmentItem } from "./AttachmentItem.vue";
 export { default as AttachmentList } from "./AttachmentList.vue";
 export { default as Autocomplete } from "./Autocomplete.vue";
-export { default as CommunicationArea } from "./CommunicationArea.vue";
+export { default as CommunicationArea } from "./communication-area/CommunicationArea.vue";
 export { default as FadedScrollableDiv } from "./FadedScrollableDiv.vue";
 export { default as AutocompleteNew } from "./frappe-ui/Autocomplete.vue";
 export { default as Link } from "./frappe-ui/Link.vue";
